[#50] DescriptionView 관련 파일 정리
DescriptionSingleView, DescriptionPageView 만 남기고 정리함.
